Added much embedded help + connections zones
I was able to add a large number of examples with sample output to ALL commands, additionally the kit now looks for Storage in the /redfish/v1/storage location, additionally it looks for drives in chassis, and can follow the storage/endpoints to the proper repository of fabric. I have also added Connections to replace StorageGroups as well as Zones to replace EndpointGroups.